refactor(main): extract CLI command dispatchers into commands/ (step 3)
Step 3 of #127. Move the per-CLI-command code out of src/main.rs
into focused command modules:

  src/commands/utils.rs         update_counters, create_json_summary,
                                print_dry_run_notice
  src/commands/max_amplitude.rs cmd_max_amplitude
  src/commands/tags.rs          CheckTagInfo, cmd_delete_tags, cmd_check_tags
  src/commands/apply.rs         cmd_apply, cmd_apply_channel
  src/commands/info.rs          cmd_info
  src/commands/undo.rs          cmd_undo
  src/commands/replaygain.rs    cmd_track_gain, cmd_album_gain

main.rs shrinks from 1259 → 114 lines: it now contains only main(),
run(), get_filename, and module declarations. Step 4 will move
run() into commands/mod.rs and reduce main.rs further.

No public library API change. CLI behavior unchanged.
