+++ src/commands/mod.rs
@@ -1,0 +1,7 @@
+pub mod apply;
+pub mod info;
+pub mod max_amplitude;
+pub mod replaygain;
+pub mod tags;
+pub mod undo;
+pub mod utils;
